+++ packages/__mocks__/@testing-library/react.tsx
@@ -3,7 +3,7 @@ import { createTheme, ThemeProvider } from '@mui/material/styles'
 import React from 'react'
 
 const react = jest.requireActual('@testing-library/react')
-const render = args => {
+const render = (args: React.ReactNode) => {
   return react.render(
     <ThemeProvider theme={createTheme()}>{args}</ThemeProvider>,
   )
