+++ products/jbrowse-web/src/tests/__image_snapshots__/alignments-test-tsx-opens-an-alignments-track-and-clicks-feature-2-snap.png
